Cleaned up `useBlockNote`, menu & toolbar components
