+++ packages/react/src/BlockNoteView.tsx
@@ -2,10 +2,10 @@ import { BlockNoteEditor, BlockSchema } from "@blocknote/core";
 import { MantineProvider } from "@mantine/core";
 import { EditorContent } from "@tiptap/react";
 import { getBlockNoteTheme } from "./BlockNoteTheme";
-import { SideMenu } from "./BlockSideMenu/components/BlockSideMenu";
-import { FormattingToolbar } from "./FormattingToolbar/components/FormattingToolbar";
-import { HyperlinkToolbar } from "./HyperlinkToolbar/components/HyperlinkToolbar";
-import { SlashMenu } from "./SlashMenu/components/SlashMenu";
+import { SideMenuWrapper } from "./SideMenu/components/SideMenuWrapper";
+import { FormattingToolbarWrapper } from "./FormattingToolbar/components/FormattingToolbarWrapper";
+import { HyperlinkToolbarWrapper } from "./HyperlinkToolbar/components/HyperlinkToolbarWrapper";
+import { SlashMenuWrapper } from "./SlashMenu/components/SlashMenuWrapper";
 
 export function BlockNoteView<BSchema extends BlockSchema>(props: {
   editor: BlockNoteEditor<BSchema>;
@@ -13,10 +13,10 @@ export function BlockNoteView<BSchema extends BlockSchema>(props: {
   return (
     <MantineProvider theme={getBlockNoteTheme()}>
       <EditorContent editor={props.editor?._tiptapEditor}>
-        <SideMenu editor={props.editor!} />
-        <FormattingToolbar editor={props.editor!} />
-        <HyperlinkToolbar editor={props.editor!} />
-        <SlashMenu editor={props.editor!} />
+        <SideMenuWrapper editor={props.editor!} />
+        <FormattingToolbarWrapper editor={props.editor!} />
+        <HyperlinkToolbarWrapper editor={props.editor!} />
+        <SlashMenuWrapper editor={props.editor!} />
       </EditorContent>
     </MantineProvider>
   );
