+++ packages/react/src/SideMenu/components/DefaultButtons/RemoveBlockButton.tsx
@@ -1,8 +1,8 @@
 import { ReactNode } from "react";
 import { BlockSchema } from "@blocknote/core";
 
-import { DragHandleMenuProps } from "../DragHandleMenu";
-import { DragHandleMenuItem } from "../DragHandleMenuItem";
+import { DragHandleMenuProps } from "../DragHandleMenu/DragHandleMenu";
+import { DragHandleMenuItem } from "../DragHandleMenu/DragHandleMenuItem";
 
 export const RemoveBlockButton = <BSchema extends BlockSchema>(
   props: DragHandleMenuProps<BSchema> & { children: ReactNode }
