+++ packages/react/src/SideMenu/components/SideMenuWrapper.tsx
@@ -1,71 +1,16 @@
+import { useEffect, useMemo, useRef, useState } from "react";
+import Tippy from "@tippyjs/react";
 import {
-  BaseUiElementCallbacks,
-  BaseUiElementState,
   Block,
   BlockNoteEditor,
   BlockSchema,
   createSideMenu,
   SideMenuCallbacks,
-  SideMenuState,
 } from "@blocknote/core";
-import { ActionIcon, Group, Menu } from "@mantine/core";
-import Tippy from "@tippyjs/react";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { AiOutlinePlus } from "react-icons/ai";
-import { MdDragIndicator } from "react-icons/md";
-import { DefaultDragHandleMenu } from "./DefaultDragHandleMenu";
 
-export const BlockSideMenuOld = <BSchema extends BlockSchema>(
-  props: Omit<SideMenuCallbacks, keyof BaseUiElementCallbacks> &
-    Omit<SideMenuState<BSchema>, keyof BaseUiElementState> & {
-      editor: BlockNoteEditor<BSchema>;
-    }
-) => {
-  const dragHandleRef = useRef<HTMLDivElement>(null);
+import { DefaultSideMenu } from "./DefaultSideMenu";
 
-  useEffect(() => {
-    const dragHandle = dragHandleRef.current;
-
-    if (dragHandle instanceof HTMLDivElement) {
-      dragHandle.addEventListener("dragstart", props.blockDragStart);
-      dragHandle.addEventListener("dragend", props.blockDragEnd);
-
-      return () => {
-        dragHandle.removeEventListener("dragstart", props.blockDragStart);
-        dragHandle.removeEventListener("dragend", props.blockDragEnd);
-      };
-    }
-
-    return;
-  }, [props.blockDragEnd, props.blockDragStart]);
-
-  const DragHandleMenu = DefaultDragHandleMenu;
-
-  return (
-    <Group spacing={0}>
-      <ActionIcon size={24} data-test={"dragHandleAdd"}>
-        <AiOutlinePlus size={24} onClick={props.addBlock} />
-      </ActionIcon>
-      <Menu
-        trigger={"click"}
-        onOpen={props.freezeMenu}
-        onClose={props.unfreezeMenu}
-        width={100}
-        position={"left"}>
-        <Menu.Target>
-          <div draggable="true" ref={dragHandleRef}>
-            <ActionIcon size={24} data-test={"dragHandle"}>
-              {<MdDragIndicator size={24} />}
-            </ActionIcon>
-          </div>
-        </Menu.Target>
-        <DragHandleMenu editor={props.editor} block={props.block} />
-      </Menu>
-    </Group>
-  );
-};
-
-export const SideMenu = <BSchema extends BlockSchema>(props: {
+export const SideMenuWrapper = <BSchema extends BlockSchema>(props: {
   editor: BlockNoteEditor<BSchema>;
 }) => {
   const [show, setShow] = useState<boolean>(false);
@@ -99,7 +44,7 @@ export const SideMenu = <BSchema extends BlockSchema>(props: {
     }
 
     return (
-      <BlockSideMenuOld
+      <DefaultSideMenu
         block={block}
         editor={props.editor}
         blockDragStart={callbacks.current.blockDragStart}
