+++ packages/react/src/index.ts
@@ -1,22 +1,22 @@
 // TODO: review directories
 export * from "./BlockNoteView";
-export * from "./BlockSideMenu/components/DefaultButtons/BlockColorsButton";
-export * from "./BlockSideMenu/components/DefaultButtons/RemoveBlockButton";
-export * from "./BlockSideMenu/components/DragHandleMenu";
-export * from "./BlockSideMenu/components/DragHandleMenuItem";
+export * from "./SideMenu/components/DefaultButtons/BlockColorsButton";
+export * from "./SideMenu/components/DefaultButtons/RemoveBlockButton";
+export * from "./SideMenu/components/DragHandleMenu/DragHandleMenu";
+export * from "./SideMenu/components/DragHandleMenu/DragHandleMenuItem";
 export * from "./FormattingToolbar/components/DefaultButtons/ColorStyleButton";
 export * from "./FormattingToolbar/components/DefaultButtons/CreateLinkButton";
 export * from "./FormattingToolbar/components/DefaultButtons/NestBlockButtons";
 export * from "./FormattingToolbar/components/DefaultButtons/TextAlignButton";
 export * from "./FormattingToolbar/components/DefaultButtons/ToggledStyleButton";
 export * from "./FormattingToolbar/components/DefaultDropdowns/BlockTypeDropdown";
-export * from "./FormattingToolbar/components/FormattingToolbar";
+export * from "./FormattingToolbar/components/FormattingToolbarWrapper";
 export * from "./ReactBlockSpec";
 export * from "./SharedComponents/Toolbar/components/Toolbar";
 export * from "./SharedComponents/Toolbar/components/ToolbarButton";
 export * from "./SharedComponents/Toolbar/components/ToolbarDropdown";
 export * from "./SlashMenu/ReactSlashMenuItem";
-export * from "./SlashMenu/components/SlashMenu";
+export * from "./SlashMenu/components/SlashMenuWrapper";
 export * from "./SlashMenu/defaultReactSlashMenuItems";
 export * from "./hooks/useBlockNote";
 export * from "./hooks/useEditorForceUpdate";
